Add fixture helpers for building queue
